Update: LIFF.initの認証処理を扉ページのhomeからアプリ全体に変更
