+++ src/index.tsx
@@ -1,8 +1,7 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import { render } from 'react-dom';
 
 // router
-import { Routes } from './routes';
+import { Routes } from './routes/routes';
 
 // redux
 import { Provider } from 'react-redux';
@@ -11,7 +10,7 @@ import { store } from './store/index';
 // asset
 import './assets/style/index.scss';
 
-ReactDOM.render(
+render(
   <Provider store={store}>
     <Routes />
   </Provider>,
