+++ src/routes/private.tsx
@@ -4,13 +4,13 @@ import { FC } from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 
 // pages
-import { Home } from './pages/home';
-import { Sub } from './pages/sub';
+import { Home } from '../pages/home';
+import { Sub } from '../pages/sub';
 
 // layout
-import { Dialog } from './layout/dialog';
+import { Dialog } from '../layout/dialog';
 
-export const Routes: FC = () => {
+export const Private: FC = () => {
   return (
     <BrowserRouter>
       <Switch>
